Improve request status handling
Updated RequestInfo objects to remove their internal
status object, and relocated the status tracking to
the sender class. This provides a cleaner API and is
reducing redundant passing of the request info object.

The sender class processRequest method now returns the
resulting status which can be accessed by the calling
method directly.