Improve CanaryConnectorConfig logs
Improved the logging for exceptions in the
CanaryConnectorConfig class by switching to
the methods provided by the abstract class.

This includes improved log output which includes
the default value used, where applicable. Critical
errors typically do not include their default
value, as it is irrelevant to the function of
the connector and simply used as a comparison to
determine whether to warn the user and shut down.