Move config load checking to abstract
Closes #41.

Moved the configuration file load checking to the
abstract level. This allows for better consistency
in the future by requiring the corresponding method
#checkRequiredConfigLoaded to be implemented, and
ensures that the check is performed and accordingly
results in a connector shutdown if required.

The relocated log has the message regarding connector
shutdown removed, as a shutdown log will already be
output as a result of the initialization failure.