Remove redundant connector config field
Removed a redundant connector configuration field from the
CanaryApiRequestBuilder.java class. There is no reason to
have a redundant/duplicate copy of the connector config as
one is already publicly accessible via the connector main.

This also resolves a bug which could be caused by the config
not being initialized when the connector shuts down, such as
a required field being missing.