Removed troublesome link in _download partial
Removed a link referencing the configuration page in a note,
as this partial doc is imported in multiple locations with
different directory levels, causing the link to fail in some
scenarios.

The quick start guide was restructured so that it includes
the configuration file information, thus this note, while
it may be useful, is also redundant and does not require an
immediate replacement in light of this removal.