Updated CanaryApiResponseStatus.java
Updated the CanaryApiResponseStatus enum-like class to
include status strings for both 'Error' and 'Errors' in
responses. getStatusFromString(String) was updated
accordingly as well. Additionally, the enum-value
ERROR_WAIT_FOR_EXPIRE was added, which is used when
messages with an error status contain a message
indicating a condition where the connector should wait
for session expiration.