Removed redundant response body logging
Removed the redundant logging of the message body
for every API response received. Extensive error
checking and parsing capabilities are present in
the Canary connector already, and this log output
results in duplicate information (parsed and cleaned
up) which can be confusing or make an error look more
serious or frequent.