Improved API response parsing
Improved parsing of API responses by skipping further
parsing (user token and/or session token) if a non-good
response is detected. This prevents errors from being
logged multiple times. In the event of an exception,
the message status is updated to UNKNOWN_ERROR now,
instead of the previous UNKNOWN_STATUS value. This
better aligns with the fact that an error occurred
while parsing response messages, and is not necessarily
related to the status being unknown or irretrievable.