Remove hardcoded request retry
Removed a hardcoded request retry step in the
CanaryApiRequestSender.java class. While there is
significant progress to be made in fully
implementing a retry with backoff, payloads are
automatically retried because they are not removed
from the payload manager if a request fails. This
hardcoded retry is likely to result in the exact
same failure if immediately retried, and can even
appear like a double logging bug.