Change how SessionManager tracks session updates
Previously, session expiration updates were triggered when session
requests were made, but these requests could fail. This would result
in the session being invalid, but the expiration tracker being updated.
Now they are made when there is a successful response to get/update
session, or when there is a non-failure request not related to session
management.